Removes old file hashing primitives and implements newer primitives. Also implements hash retrieval function.

+++ fs_monitor/src/state.rs
@@ -18,5 +18,5 @@ pub enum FileState {
 pub struct FileInfo {
     pub file_type: FileType,
     pub state: FileState,
-    pub version: i32,
+    pub version: usize,
 }
